Refactor publication card components
